switch to better data types
refactor

update toolchain, use more of nix-shell, remove old toolchain, move run script to makefile

Clean up a bit, add README, remove old run script

fix typo

+++ include/kheap.h
@@ -1,13 +1,10 @@
-// kheap.h -- Interface for kernel heap functions, also provides
-//            a placement malloc() for use before the heap is 
-//            initialised.
-//            Written for JamesM's kernel development tutorials.
-
 #ifndef KHEAP_H
 #define KHEAP_H
 
-#include "common.h"
-#include "ordered_array.h"
+#include <stdint.h>
+#include <stddef.h>
+#include <common.h>
+#include <ordered_array.h>
 
 #define KHEAP_START         0xC0000000
 #define KHEAP_INITIAL_SIZE  0x100000
@@ -67,30 +64,30 @@ void free(void *p, heap_t *heap);
    parameter representations are available in kmalloc, kmalloc_a,
    kmalloc_ap, kmalloc_p.
 **/
-uint32_t kmalloc_int(uint32_t sz, int align, uint32_t *phys);
+uint32_t kmalloc_int(size_t sz, int align, uint32_t *phys);
 
 /**
    Allocate a chunk of memory, sz in size. The chunk must be
    page aligned.
 **/
-uint32_t kmalloc_a(uint32_t sz);
+uint32_t kmalloc_a(size_t sz);
 
 /**
    Allocate a chunk of memory, sz in size. The physical address
    is returned in phys. Phys MUST be a valid pointer to uint32_t!
 **/
-uint32_t kmalloc_p(uint32_t sz, uint32_t *phys);
+uint32_t kmalloc_p(size_t sz, uint32_t *phys);
 
 /**
    Allocate a chunk of memory, sz in size. The physical address 
    is returned in phys. It must be page-aligned.
 **/
-uint32_t kmalloc_ap(uint32_t sz, uint32_t *phys);
+uint32_t kmalloc_ap(size_t sz, uint32_t *phys);
 
 /**
    General allocation function.
 **/
-uint32_t kmalloc(uint32_t sz);
+uint32_t kmalloc(size_t sz);
 
 /**
    General deallocation function.
